feat(all): Tests built and basic formatted output completed
1. Set up kernel unit tests with catch2

2. Complete the basic string formatting function, similar to std::format

3. Adjust the cmake script, set different compilation parameters for different
targets (kernel, test), and control whether to test or compile kernel according
to BUILD_TEST

4. The previous kernel entry address was 0x80000, which belonged to the ROM
area, so we could not use the bss and data fields. Now we reset the kernel
entry address to 0x40000000 and set the kernel address space to 128M. Now we
can use static/global variables.

5. Move types.h to util

Signed-off-by: littleWhitePoplar <1016164355@qq.com>